+++ test/clj_kondo/main_test.clj
@@ -1254,7 +1254,8 @@ foo/foo ;; this does use the private var
     Number
   IFoo
   (-foo [this]
-    (recur this)))"))))
+    (recur this)))")))
+  (is (empty? (lint! "(defrecord Foo [] clojure.lang.ILookup (valAt [_ _] (letfn [(_foo [_x] (recur 1))])))"))))
 
 (deftest lint-as-test
   (assert-submaps
@@ -1782,9 +1783,6 @@ foo/foo ;; this does use the private var
                      {:linters {:unresolved-symbol {:level :error}}})))
   (is (empty? (lint! (io/file "corpus" "core.rrb-vector.clj")
                      {:linters {:unresolved-symbol {:level :error}}})))
-  ;; don't crash in this example. maybe in the future have better syntax checking for ns
-  ;; see GH-497
-  (is (empty? (lint! "(ns circleci.rollcage.test-core (:require [clojure.test :refer :refer [deftest]]))")))
   (is (empty? (lint! "(defn get-email [{email :email :as user :or {email user}}] email)"
                      {:linters {:unresolved-symbol {:level :error}}})))
   (is (empty? (lint! "(ns repro (:require [clojure.string :refer [starts-with?]]))