+++ CHANGELOG.md
@@ -5,6 +5,9 @@ For a list of breaking changes, check [here](#breaking-changes).
 ## Unreleased
 
 - [#1794](https://github.com/clj-kondo/clj-kondo/issues/1794): Add a linter for line length
+- [#1460](https://github.com/clj-kondo/clj-kondo/issues/1460): Add a linter for unknown `:require` options
+- [#1807](https://github.com/clj-kondo/clj-kondo/issues/1807): false positive with map transducer in cljs
+- [#1806](https://github.com/clj-kondo/clj-kondo/issues/1806): false positive recur mismatch with letfn
 - [#1800](https://github.com/clj-kondo/clj-kondo/issues/1800): Add `:existing-alias`, a linter that checks qualified symbols and keywords to see if they're using an existing alias.
 
 ## 2022.09.08